Turned the searchable selector into an HOC.
